Added solution - LeetHub

+++ 0 - 1 Knapsack Problem - GFG/0-1-knapsack-problem.cpp
@@ -10,22 +10,19 @@ class Solution
     //Function to return max value that can be put in knapsack of capacity W.
     int knapSack(int w, int wt[], int val[], int n) 
     { 
-       vector<vector<long long int>> dp(n+1,vector<long long int>(w+1,0));
-       
-       for(int i=1;i<=n;i++)
-       {
-           for(int j=1;j<=w;j++)
-           {
-               if(wt[i-1]<=j)
-               {
-                   dp[i][j] = max(val[i-1]+dp[i-1][j-wt[i-1]],dp[i-1][j]);
-               }
-               else dp[i][j] = dp[i-1][j];
-           }
-       }
-       
-       return dp[n][w];
-       
+        vector<vector<int>>dp(n+1,vector<int>(w+1,0));
+        
+        for(int i=1;i<=n;i++)
+        {
+            for(int j=1;j<=w;j++)
+            {
+                if(wt[i-1]<=j)
+                {
+                    dp[i][j] = max(val[i-1]+dp[i-1][j-wt[i-1]],dp[i-1][j]);
+                }
+                else dp[i][j] = dp[i-1][j];
+            }
+        }return dp[n][w];
     }
 };
